jungah - 영수증 인증 후 이미지파일 삭제 시 속도 저하 해결 (그나마 ㅠㅠ)

+++ src/main/java/com/project/foodiefoodie/proBoard/domain/ProBoard.java
@@ -26,7 +26,7 @@ public class ProBoard extends StoreTimeDTO {
     private String content;
     private String title;
     private String hashTag;
-    private float avgStarRate;
+    private double avgStarRate;
     private int review_cnt;
     private int report_cnt;
